Issue 128 work: - Auto-scrolling of rule list implemented. - Expanded/collapsed indicators added for groups in the masterlist, other than the lowest level of groups (whose indicators are hidden when the root node is hidden, bah humbug).
git-svn-id: http://better-oblivion-sorting-software.googlecode.com/svn@3605 bfee223e-1fc4-11de-b650-e715bd6d7cf1

+++ code/boss-gui/trunk/GUI/UserRuleEditor.h
@@ -81,7 +81,7 @@ public:
 private:
 	void ReDrawRuleList();					//Empties the RuleListScroller and then re-populates it with RuleBoxClass objects for the rules in the RuleList object.
 	RuleList userlist;
-	uint32_t selectedRuleIndex;
+	size_t selectedRuleIndex;
 	wxScrolled<wxPanel> *RuleListScroller;
 };
